chore(tests): move shared settings out of conftest modules
Importing from a conftest.py is unsafe: pytest loads conftest modules
specially, so a direct import can yield a duplicated/mismatched module.
Declare the Settings dataclasses in plain settings.py helpers next to
each broker's tests and import them from both the conftest and the test
files.

Also move SettingsCluster and the settings_cluster fixture down into
tests/brokers/redis/cluster/, where they are the only users.
