+++ tests/brokers/confluent/test_connect.py
@@ -5,7 +5,7 @@ from faststream.confluent import KafkaBroker
 from faststream.confluent.helpers import config
 from tests.brokers.base.connection import BrokerConnectionTestcase
 
-from .conftest import Settings
+from .settings import Settings
 
 
 @pytest.mark.connected()
