+++ tests/brokers/kafka/conftest.py
@@ -1,13 +1,8 @@
-from dataclasses import dataclass
-
 import pytest
 
 from faststream.kafka import KafkaRouter
 
-
-@dataclass
-class Settings:
-    url: str = "localhost:9092"
+from .settings import Settings
 
 
 @pytest.fixture(scope="session")
