+++ tests/brokers/kafka/settings.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Settings:
+    url: str = "localhost:9092"
