+++ tests/brokers/mqtt/settings.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Settings:
+    host: str = "localhost"
+    port: int = 1883
