+++ tests/brokers/nats/settings.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Settings:
+    url: str = "nats://localhost:4222"
