+++ tests/brokers/redis/cluster/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from .settings import SettingsCluster
+
+
+@pytest.fixture(scope="session")
+def settings_cluster():
+    return SettingsCluster()
