+++ tests/brokers/redis/cluster/test_cluster_connect.py
@@ -13,7 +13,8 @@ import pytest
 
 from faststream.redis import RedisClusterBroker, StreamSub
 from tests.brokers.base.connection import BrokerConnectionTestcase
-from tests.brokers.redis.conftest import SettingsCluster
+
+from .settings import SettingsCluster
 
 
 @pytest.mark.connected()
