+++ tests/brokers/redis/cluster/test_cluster_pubsub_more.py
@@ -5,7 +5,8 @@ import pytest
 
 from faststream.redis import RedisClusterBroker
 from tests.brokers.redis.basic import RedisClusterTestcaseConfig
-from tests.brokers.redis.conftest import SettingsCluster
+
+from .settings import SettingsCluster
 
 
 @pytest.mark.slow()
