+++ tests/brokers/redis/settings.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Settings:
+    url: str = "redis://localhost:6379"
+    host: str = "localhost"
+    port: int = 6379
